+++ internal/transport/http/handler/get_original.go
@@ -1,4 +1,4 @@
-package http
+package handler
 
 import (
 	"net/http"
