+++ internal/transport/http/handler/get_short.go
@@ -1,4 +1,4 @@
-package http
+package handler
 
 import (
 	"net/http"
