+++ internal/transport/http/handler/redirect.go
@@ -1,4 +1,4 @@
-package http
+package handler
 
 import (
 	"errors"
