+++ internal/transport/http/handler/shorten.go
@@ -1,4 +1,4 @@
-package http
+package handler
 
 import (
 	"net/http"
